+++ lib/constants.ts
@@ -1,6 +1,9 @@
 export const TASK_APPEND_MESSAGES =
   "cypress-cucumber-preprocessor:append-messages";
 
+export const TASK_TEST_CASE_STARTED =
+  "cypress-cucumber-preprocessor:test-case-started";
+
 export const TASK_TEST_STEP_STARTED =
   "cypress-cucumber-preprocessor:test-step-started";